fix 10.5 MISRA VIOLATION

+++ SERVICE/Heat_CTRL.c
@@ -1,7 +1,6 @@
 #define F_CPU 4000000
 
 #include "../STD_Types.h"
-#include "../macros.h"
 
 #include "../HAL/TC72_Driver.h"
 #include "../MCAL/TIMER0_config.h"
@@ -19,7 +18,6 @@ void Heat_CNRL_Duty_Cycle(u8 Current_Temp, u8 Set_Temp)
 	
 	Temp_Volt = (( (f32)Set_Temp - (f32)Current_Temp) / (f32)100.0 ) * (f32)10.0 ;
 	Potontiometer_Volt  = ADC_OS_GetValue();
-	//Potontiometer_Volt = 5;
 	Duty_value = ((( Potontiometer_Volt * (f32)2.0 ) / (f32)10.0 ) * Temp_Volt ) / (f32)10.0;
 	PWM_dutyCycle= Heat_CNRL_Duty_Value_Map(Duty_value);
 	TIMER0_void_fastPWM(PWM_dutyCycle);
@@ -29,7 +27,7 @@ void Heat_CNRL_Duty_Cycle(u8 Current_Temp, u8 Set_Temp)
 
 u8 Heat_CNRL_Duty_Value_Map(f32 Duty_Value)
 {
-	u8 PWM_duty = (u8)((u8)255 - (u8)(Duty_Value * (u8)255));
+	u8 PWM_duty = (u8)((u8)255 - (u8)((u8)Duty_Value * (u8)255));
 	return PWM_duty;
 	
 }
